move all constants to one file
