+++ js/utils.js
@@ -1,9 +1,4 @@
-const ERRORS = {
-  UNIQUE_ERROR: 'Один и тот же хэш-тег не может быть использован дважды!',
-  MAX_COUNT_ERROR: 'Нельзя указать больше пяти хэш-тегов!',
-  CORRECT_TITLE_ERROR: 'Неправильный формат хэш-тега!',
-  MAX_LENGTH_ERROR: 'Длина комментария не может составлять больше 140 символов!',
-};
+
 const getRandomNumber = (min, max) => {
   min = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   max = Math.floor(Math.max(Math.abs(min),Math.abs(max)));
@@ -17,4 +12,4 @@ const getRandomNumberExcept = (min, max, canceledNum) => {
 const isEscapeKey = (evt) => evt.key === 'Escape';
 const findMiniatureData = (miniatures, src) => miniatures.find((item) => src.endsWith(item.url));
 
-export {getRandomNumber, getRandomNumberExcept, checkStrFitsMaxLength, isEscapeKey, findMiniatureData, ERRORS};
+export {getRandomNumber, getRandomNumberExcept, checkStrFitsMaxLength, isEscapeKey, findMiniatureData};
